Add files via upload
An attempt at HW1P2 that crosses the low cutoff, reaching 74% accuracy on the verification data, by means of an MLP with a single hidden layer. I will continue to experiment with ablations and post updates should I cross, say, 80% and 90% cutoffs.

If you wish to run the notebook for yourself, I find that it is best run in a kaggle notebook (trying in Google Colab I frequently encounter out of memory errors, even upon decreasing num_workers).